Add files via upload
These are the codes what I have for web-scraping so far. So far, I am just scraping text from financial announcements in Chinese. The next step will be to determine the overall emotion of each text. The ultimate goal is to deduce a relationship between emotion (of a firm's announcement) and change in its stock price.