Add files via upload
This is the codes that I wrote for my research in biostatistics. Here's the context. We want to examine how the possibility of getting a certain disease is affected by its potential risk factor plus among some patients, but due to budget constraint, it was infeasible to take accurate measurements on this risk factor for every patient, so we had to settle for an approximate measurement for most. Yet we still want to make valid conclusion about the relationship between the disease occurrence and this risk factor.
To do so, I first reformatted the problem as follows: “If all measurements were accurate, it would have been a straightforward problem which I knew how to tackle. Now that there are these twists and turns in my data, what twists and turns should I add to my methodology to ensure valid conclusion from the data?” 
Then I discussed and learned from my supervisor that this problem could be solved by adding weights to the data while using our usual estimation method. In other words, a weighted estimation (like a weighted average). Specifically, we first came up with an initial set of estimates as if the data were all accurate. Then we used that estimate to get weights for all data, and got a new set of estimates, then updated the weights, and repeated the process until the current and previous sets of estimates were in close range.
Then, in terms of actual coding, I first used prespecified values to generate 2000 sets of mock data (there were 2 sets of prespecified values, each with 3 varying degree of approximate measurements ranging from 50% to 90%), then I coded the estimation algorithm and tested it on the mock data. When I calculated the average of the estimates, I found them mostly in close range with the original values (by that I mean within 10^-2). That’s when I knew I have come up with an algorithm for problems like this.