Add files via upload
Disclaimer: The codes aren't mine.
For this project, we want to examine how reliable the forecasts are for autoregressive moving-average, ARMA(p,q) and for autoregressive conditional-variance series, ARCH(1) in this case ( approximated by ARMA(0, 0) ). Here, p = # of parameters under AR and q = # of parameters under MA.
We measured the mean-squared error (MSE) of 2-, 5- and 10-step predictions for forecasting each of the following: ARMA(1,1), ARMA(1,0), an ARCH(1). We found out that, if the model's complex like ARMA(1,1), even large samples (100 for training and 10 for testing) would help little for 5- and 10-step predictions because, when the forecast's this far out, it must rely on previous predicted values too. Once the estimation has to rely on previous estimates, the variance simply adds up. On the other hand, if the model is simple like ARMA(1,0), even small sample (20 for training and 10 for testing) gives pretty accurate predictions for steps far out like 5- and 10-step. As for ARCH(1), we "pretended" that the generated data came from an ARMA series and used AIC to select p and q, hence got ARMA(0,0). The performance of predictions was surprisingly well.